% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_version.R
\name{check_r_version}
\alias{check_r_version}
\title{Check R version}
\usage{
check_r_version(version = NULL)
}
\arguments{
\item{version}{Something to coerce into an R version.  Valid
values are \code{NULL} (the current session's R version), a
character string of the form \code{x.y.z}, a character string
\code{oldrel} or \code{release} or a \code{numeric_version}
object.}
}
\description{
Utility function for checking an R version
}
