% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_files}
\alias{download_files}
\title{Download multiple files}
\usage{
download_files(urls, dest_dir, ..., labels = NULL, overwrite = FALSE,
  count = TRUE, dest_file = NULL, copy_file_urls = TRUE,
  progress = NULL, report = TRUE)
}
\arguments{
\item{urls}{A character vector of urls}

\item{dest_dir}{A single existing directory to download files into}

\item{...}{Currently ignored}

\item{labels}{A character vector of labels to use to describe the
files being downloaded when printing (defaults to
\code{basename(urls)})}

\item{overwrite}{Overwrite files that exist already?  If
\code{FALSE} (the default) then existing files are skipped.}

\item{count}{Logical, indicating if a count of progress across the
urls should be included.}

\item{dest_file}{If the files should be renamed as they are
downloaded, include a vector of filenames here the same length
as \code{urls}.  Directory components will be created, within
\code{dest_dir}.}

\item{copy_file_urls}{Logical, indicating if \code{file:///} urls
should be copied into \code{dest_dir}}

\item{progress}{Print a progress bar?}

\item{report}{Print a summary?}
}
\value{
A character vector, the same length as \code{urls}, with
  the destination file paths (even if no downloading was done).  A
  failure to download a file (e.g., a 403 forbidden, 404 not
  found, or general network error) will result in an R error.
}
\description{
Download multiple files.
}
