% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_sources.R
\name{package_sources}
\alias{package_sources}
\title{Collect information on package sources}
\usage{
package_sources(cran = NULL, repos = NULL, github = NULL, local = NULL,
  expire = NULL)
}
\arguments{
\item{cran}{A single URL for the CRAN repo to use.  If
\code{NULL}, then this will use the rstudio mirror.}

\item{repos}{A character vector of additional repositories to use.
Repositories of the form \code{drat://username} will be
translated to a drat repository hosted on github (e.g.,
\code{https://username.github.io/drat})}

\item{github}{A vector of github package specifications (e.g.,
username/password).  The full syntax as implemented in
\code{\link{parse_github_repo_spec}} is supported (packages in
subdirectories, or referencing a particular branch, commit or
tag).}

\item{local}{A character vector of local files to include.  Can be
directories or built packages (source or binary)}
}
\description{
Collect information on package sources
}

