% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provision_dependencies.R
\name{provision_dependencies}
\alias{provision_dependencies}
\alias{provision_dependencies_bootstrap}
\title{Provision dependencies for a package}
\usage{
provision_dependencies(lib, path_description = ".", ..., src = NULL,
  read_travis = FALSE)

provision_dependencies_bootstrap(lib = ".packages", src = NULL,
  read_travis = FALSE, strict_lib = TRUE)
}
\arguments{
\item{lib}{Library to provision}

\item{path_description}{Path to the \code{DESCRIPTION} file}

\item{...}{Additional arguments to \code{\link{provision_library}}}

\item{src}{An optional description of additional packages, using
\code{\link{package_sources}}.}

\item{read_travis}{Logical, indicating if the \code{.travis.yml}
should be read (if present).  If \code{TRUE} then packages
listed within \code{r_github_packages} will be used in
provisioning.}

\item{strict_lib}{Should the bootstrap script install provisionr
within the local library too?}
}
\description{
Provision dependencies for a package.  The
\code{provision_dependencies_bootstrap} function writes out a
bootstrap script.
}
