% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_level.R
\name{provision_library}
\alias{provision_library}
\title{Create or update a library}
\usage{
provision_library(packages, lib, platform = NULL, version = NULL,
  src = NULL, path_drat = NULL, check_dependencies = TRUE,
  installed_action = "replace")
}
\arguments{
\item{packages}{A character vector of packages to include}

\item{lib}{A path to the library; if it does not exist, it will be
created.}

\item{platform}{The platform to create the library for.  If
\code{NULL} then we build for the current platform (using
\code{\link{install_packages}} if \code{version} is \code{NULL}
or compatible with our current version).  Otherwise this can be
one of \code{"windows"}, \code{"macosx"},
\code{"macosx/mavericks"} or \code{"linux"}, correspinding to
the different directories that binaries live in (in the case of
\code{"linux"} there are no binaries and things are a little
more complicated).}

\item{version}{The version of R to install packages for.  By
default, we use the same version (major.minor) as the current
running R version.  Otherwise provide the desired version number
as a string or \code{numeric_version} object.}

\item{src}{An optional description of additional packages, using
\code{\link{package_sources}} (can be either a
\code{package_sources} object or a \code{local_drat} object).}

\item{path_drat}{When using a \code{package_sources} object for
\code{src}, the path to store the drat repository.  If not
given, then a temporary directory will be used.}

\item{check_dependencies}{Logical, indicating if dependencies of
\code{packges} should be checked.  If \code{TRUE}, then any
missing dependencies (packages in Depends, Imports or LinkingTo
of the requested packages) will be installed.}

\item{installed_action}{The behaviour when some packages are
already installed.  Options are \code{"replace"} (will
re-install the package and, for cross-installation, its
dependencies), \code{"upgrade_all"} (upgrade all packages that
lag behind the versions in repositories), \code{"upgrade"}
(upgrade packages listed in \code{"packages"} only, but not
their dependencies) and \code{"skip"} (do not install or upgrade
any package that is already installed).}
}
\description{
Create or update a library of packages.
}
\details{
Cross installation of binary files is difficult and I need to come
up with a way of making that work nicely.
}
\author{
Rich FitzJohn
}

