% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{download_cran}
\alias{download_cran}
\title{Download some of CRAN}
\usage{
download_cran(packages, path, r_version = NULL, target = "windows",
  suggests = FALSE, package_sources = NULL, missing_index_is_error = TRUE,
  progress = NULL)
}
\arguments{
\item{packages}{Character vector of packages to download}

\item{path}{Destination}

\item{r_version}{Target R version}

\item{target}{Target platform (use "ALL" for all platforms)}

\item{suggests}{Include suggested packages too?}

\item{package_sources}{A \code{provisionr::package_sources} object}

\item{missing_index_is_error}{Is a missing PACKAGES index an
error?  Set this to \code{FALSE} when downloading from
incomplete repositories (e.g., a drat repo that contains only
source files if you are trying to download binaries)}

\item{progress}{Control progress bar printing}
}
\description{
Download a fraction of CRAN to serve locally.
}
